+++ src/lips.js
@@ -6,9 +6,6 @@
  *
  * build: {{DATE}}
  */
-/*
- * TODO: Pair.prototype.toObject = alist to Object
- */
 "use strict";
 /* global define, module, setTimeout, jQuery */
 (function(root, factory) {
@@ -95,24 +92,57 @@
         var special_forms = special_tokens.map(s => specials[s].name);
         var parents = 0;
         var first_value = false;
+        var specials_stack = [];
+        var single_list_specials = [];
+        function pop_join() {
+            var top = stack[stack.length - 1];
+            if (top instanceof Array && top[0] instanceof Symbol &&
+                special_forms.includes(top[0].name) &&
+                stack.length > 1) {
+                stack.pop();
+                if (stack[stack.length - 1].length === 1 &&
+                    stack[stack.length - 1][0] instanceof Symbol) {
+                    stack[stack.length - 1].push(top);
+                } else if (stack[stack.length - 1].length === 0) {
+                    stack[stack.length - 1] = top;
+                } else if (stack[stack.length - 1] instanceof Pair) {
+                    if (stack[stack.length - 1].cdr instanceof Pair) {
+                        stack[stack.length - 1] = new Pair(
+                            stack[stack.length - 1],
+                            Pair.fromArray(top)
+                        );
+                    } else {
+                        stack[stack.length - 1].cdr = Pair.fromArray(top);
+                    }
+                } else {
+                    stack[stack.length - 1].push(top);
+                }
+            }
+        }
         tokens.forEach(function(token) {
             var top = stack[stack.length - 1];
             if (special_tokens.indexOf(token) !== -1) {
                 special = token;
+                stack.push([specials[special]]);
+                if (!special) {
+                    single_list_specials = [];
+                }
+                single_list_specials.push(special);
             } else if (token === '(') {
                 first_value = true;
                 parents++;
                 if (special) {
-                    stack.push([specials[special]]);
-                    special = null;
+                    specials_stack.push(single_list_specials);
+                    single_list_specials = [];
                 }
                 stack.push([]);
+                special = null;
             } else if (token === '.' && !first_value) {
                 stack[stack.length - 1] = Pair.fromArray(top);
             } else if (token === ')') {
                 parents--;
                 if (!stack.length) {
-                    throw new Error('Unbalanced parenthesis');
+                    throw new Error('Unbalanced parenthesis 1');
                 }
                 if (stack.length === 1) {
                     result.push(stack.pop());
@@ -124,24 +154,14 @@
                     } else if (top instanceof Pair) {
                         top.append(Pair.fromArray(list));
                     }
-                    if (top instanceof Array && top[0] instanceof Symbol &&
-                        special_forms.includes(top[0].name) &&
-                        stack.length > 1) {
-                        stack.pop();
-                        if (stack[stack.length - 1].length === 0) {
-                            stack[stack.length - 1] = top;
-                        } else if (stack[stack.length - 1] instanceof Pair) {
-                            if (stack[stack.length - 1].cdr instanceof Pair) {
-                                stack[stack.length - 1] = new Pair(
-                                    stack[stack.length - 1],
-                                    Pair.fromArray(top)
-                                );
-                            } else {
-                                stack[stack.length - 1].cdr = Pair.fromArray(top);
-                            }
-                        } else {
-                            stack[stack.length - 1].push(top);
+                    if (specials_stack.length) {
+                        single_list_specials = specials_stack.pop();
+                        while (single_list_specials.length) {
+                            pop_join();
+                            single_list_specials.pop();
                         }
+                    } else {
+                        pop_join();
                     }
                 }
                 if (parents === 0 && stack.length) {
@@ -151,14 +171,21 @@
                 first_value = false;
                 var value = parse_argument(token);
                 if (special) {
-                    value = [specials[special], value];
+                    // special without list like ,foo
+                    stack[stack.length - 1][1] = value;
+                    value = stack.pop();
                     special = false;
                 }
+                top = stack[stack.length - 1];
                 if (top instanceof Pair) {
                     var node = top;
                     while (true) {
                         if (node.cdr === nil) {
-                            node.cdr = value;
+                            if (value instanceof Array) {
+                                node.cdr = Pair.fromArray(value);
+                            } else {
+                                node.cdr = value;
+                            }
                             break;
                         } else {
                             node = node.cdr;
@@ -172,7 +199,8 @@
             }
         });
         if (stack.length) {
-            throw new Error('Unbalanced parenthesis');
+            console.log({end: stack.slice()});
+            throw new Error('Unbalanced parenthesis 2');
         }
         return result.map((arg) => {
             if (arg instanceof Array) {
@@ -437,6 +465,16 @@
         this.value = value;
     }
     // ----------------------------------------------------------------------
+    // :: Unquote is used for multiple backticks and unquote
+    // ----------------------------------------------------------------------
+    function Unquote(value, count) {
+        this.value = value;
+        this.count = count;
+    }
+    Unquote.prototype.toString = function() {
+        return '<#unquote[' + this.count + '] ' + this.value + '>';
+    };
+    // ----------------------------------------------------------------------
     // :: function that return macro for let and let*
     // ----------------------------------------------------------------------
     function let_macro(asterisk) {
@@ -699,6 +737,7 @@
         }),
         quasiquote: new Macro(function(arg) {
             var self = this;
+            var max_unquote = 0;
             function recur(pair) {
                 if (pair instanceof Pair) {
                     var eval_pair;
@@ -720,23 +759,39 @@
                         }
                         return eval_pair;
                     }
-                    if (Symbol.is(pair.car, 'unquote-splicing')) {
-                        eval_pair = evaluate(pair.cdr.car, self);
-                        if (!eval_pair instanceof Pair) {
-                            throw new Error('Value of unquote-splicing' +
-                                            ' need to be pair');
-                        }
-                        return eval_pair;
-                    }
                     if (Symbol.is(pair.car, 'unquote')) {
-                        if (pair.cdr.cdr !== nil) {
-                            return new Pair(
-                                evaluate(pair.cdr.car, self),
-                                pair.cdr.cdr
+                        var head = pair.cdr;
+                        var node = head;
+                        var parent = node;
+                        var unquote_count = 1;
+                        while (Symbol.is(node.car.car, 'unquote')) {
+                            parent = node;
+                            unquote_count++;
+                            node = node.car.cdr.car;
+                        }
+                        if (unquote_count > max_unquote) {
+                            max_unquote = unquote_count;
+                        }
+                        // we use Unquote to proccess inner most unquote first
+                        // in unquote function afer processing whole s-expression
+                        if (parent === node) {
+                            if (pair.cdr.cdr !== nil) {
+                                return new Pair(
+                                    new Unquote(pair.cdr.car, unquote_count),
+                                    pair.cdr.cdr
+                                );
+                            } else {
+                                return new Unquote(pair.cdr.car, unquote_count);
+                            }
+                        } else if (parent.cdr.cdr !== nil) {
+                            parent.car.cdr = new Pair(
+                                new Unquote(node, unquote_count),
+                                parent.cdr === nil ? nil : parent.cdr.cdr
                             );
                         } else {
-                            return evaluate(pair.cdr.car, self);
+                            parent.car.cdr = new Unquote(node, unquote_count);
                         }
+                        return head.car;
                     }
                     var car = pair.car;
                     if (car instanceof Pair) {
@@ -750,7 +805,34 @@
                 }
                 return pair;
             }
-            return new Quote(recur(arg.car));
+            function unquote(pair) {
+                if (pair instanceof Unquote) {
+                    if (max_unquote === pair.count) {
+                        return evaluate(pair.value, self);
+                    } else {
+                        return new Pair(
+                            new Symbol('unquote'),
+                            new Pair(
+                                unquote(pair.value),
+                                nil
+                            )
+                        );
+                    }
+                }
+                if (pair instanceof Pair) {
+                    var car = pair.car;
+                    if (car instanceof Pair || car instanceof Unquote) {
+                        car = unquote(car);
+                    }
+                    var cdr = pair.cdr;
+                    if (cdr instanceof Pair || cdr instanceof Unquote) {
+                        cdr = unquote(cdr);
+                    }
+                    return new Pair(car, cdr);
+                }
+                return pair;
+            }
+            return new Quote(unquote(recur(arg.car)));
         }),
         clone: function(list) {
             return list.clone();
@@ -1032,6 +1114,7 @@
         }
         return result;
     }
+
     // ----------------------------------------------------------------------
     // cadr caddr cadadr etc.
     combinations(['d', 'a'], 2, 5).forEach((spec) => {
@@ -1107,6 +1190,11 @@
             return code;
         }
     }
+
+    // ----------------------------------------------------------------------
+    function exec(string, env) {
+        return parse(tokenize(string)).map((code) => evaluate(code, env));
+    }
     // ----------------------------------------------------------------------
 
     function balanced(code) {
@@ -1116,6 +1204,7 @@
         var close = parenthesis.filter(p => p === '(');
         return open.length === close.length;
     }
+
     // --------------------------------------
     Pair.unDry = function(value) {
         return new Pair(value.car, value.cdr);
@@ -1148,6 +1237,7 @@
     };
     return {
         version: '{{VER}}',
+        exec: exec,
         parse: parse,
         tokenize: tokenize,
         evaluate: evaluate,