+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Next
+### Features
+* exec api method
+### Bugfixes
+* fix processing of multiple backquote and unquote
+
 ## 0.2.1/0.2.2
 * fix build system